Migrate CPU to RV64G and pass 107/107 compliance suite (#111)
* Migrate CPU to RV64G and fix atomics datapath

Widen execution datapath, physical register file, memories, and CSRs to
64-bit XLEN. Fix several critical datapath and control bugs exposed by
the migration:
- Correct 2-element list resolution in mkOrChain to wire outputs properly
  in decoded opcode lookup tables.
- Separate word vs doubleword atomic operations in the sign-extension
  path to prevent doubleword results from being sign-extended.
- Switch memory request size selection from store buffer dequeue size to
  memory pipeline size for atomic direct writes.
- Add word-aware operand extension and comparators in AtomicUnit to
  evaluate 32-bit word min/max operations properly.
- Include SC operations in the memory size bit 1 decode logic.
- Gate atomic responses with unit busy status to block spurious writes.

All 112 generated SystemVerilog modules pass slang linter with 0 errors
and 0 warnings, as well as Yosys formal validation. 99 of 107 RISC-V 64-bit
tests pass in lock-step Spike cosimulation (rv64ui 52/52, rv64um 13/13,
rv64ua 17/19).

* Gate atomic dispatch on in-flight memory loads

Atomics previously dispatched while an earlier DMEM load was in flight
in the memory pipeline. The atomic unit would erroneously capture the
earlier load's response, prematurely write back updated data, and corrupt
memory before its own read occurred.

Gating atomic dispatch on active loads in the memory pipeline and
pending DMEM responses ensures atomics only execute after all prior loads
clear.

* Support 64-bit integer operands in FP execution

Widen integer register data routed to the FP reservation station and
execution unit to 64 bits on RV64, allowing FMV.D.X and 64-bit moves to
transfer full 64-bit values.

Bypass NaN-unboxing for FMV.X.W per the RISC-V specification, and
pause ROB retirement when replaying deferred FP commits to avoid
corrupting committed rename state.

* Implement 64-bit float/integer conversions

Add FPLongConverter circuit implementing FCVT.{L,LU}.{S,D} and
FCVT.{S,D}.{L,LU} for RV64.

Route new FPU opcodes 24-27 and 58-61 through CPU decoder LUT and
FPExecUnit_D. Update testbench suite timeout for lrsc test to allow
all 107 RV64 tests to pass cosimulation.

* Fix CI lint and update Lean library root

Regenerate lean/Shoumei/All.lean to include FPLongConverter and its
proofs, satisfying the Lean root currency CI check.

Fix format string mismatches and unused variable warnings in testbench
cosimulation drivers to satisfy cppcheck CI lint rules.

* Configure CI for 64-bit RISC-V toolchain

Update CI workflow and setup script to install and cache the 64-bit
RISC-V GCC toolchain (riscv64-elf-gcc) so that riscv-tests can compile
all RV64 compliance ELFs during Verilator and Spike simulation jobs.

* Migrate test suite to 64-bit RISC-V target

Update testbench compiler flags and test harness sources for 64-bit
RISC-V execution.

The CPU datapath was widened to RV64G, causing tests compiled for RV32
to fail due to 64-bit sign extension in subword loads, 64-bit CSR
widths, and 64-bit sign-extended immediate values from fmv.x.w.

* Fix 64-bit CLINT MMIO and CSR RVVI trace extension

Support 64-bit timer accesses in the Spike simulation oracle and align
the RVVI CSR retirement trace upper bits with physical register file
injection.

FreeRTOS on RV64 accesses mtime and mtimecmp via 64-bit stores (sd) to
the CLINT MMIO region. The previous Spike oracle implementation only
loaded and stored 32-bit words, causing mtimecmp upper bits to remain
unwritten and timer interrupts to never trigger during cosimulation.

Additionally, when CSR instructions retire in slot 1 with XLEN=64, the
RVVI trace sign-extended bit 31 instead of zero-extending the register
value to match the physical register file CDB injection, causing
register value mismatches on subsequent arithmetic instructions.

* Fix RV64 testbench interrupt handler and includes

Update testbench C tests to support 64-bit register saves in the trap
handler, read cycle counter CSR for timer advancement, and include
shoumei.h rather than missing headers.

On RV64 targets, saving caller-saved registers in a 32-bit layout caused
stack corruption during interrupts. Reading MTIME_LO from cached memory
prevented timer tick advancement, while missing test header inclusions
broke CI test compilation.

* Fix test discovery and 64-bit Spike caching in CI

Update testbench Makefile test discovery to use shell globbing and
configure 64-bit Spike caching and test compilation in CI.

GNU Make evaluated $(wildcard ...) test globs during makefile parsing,
which expanded to an empty list before target prerequisites built the
ELF binaries. When run on clean CI environments, run-suite.sh received
no input ELFs and failed with exit code 2.

In addition, the cosimulation CI job restored an outdated 32-bit Spike
binary due to a stale cache key prefix, preventing 64-bit cosimulation.

* Fix CC override in riscv-tests Makefile

Define CC using RISCV_PREFIX to prevent environment CC variables from
overriding the cross-compiler with the host compiler during CI builds.

In CI environments where CC is defined in the runner environment,
checking ifeq ($(origin CC),default) caused make to retain host gcc
rather than detecting the 64-bit RISC-V cross-compiler, failing test
ELF compilation.

* Fix cosim tohost registration and compiler flags

Configure DPI tohost and putchar addresses after DUT reset in the
cosimulation driver so that initial blocks in the SystemVerilog
testbench do not overwrite symbol addresses found in the test ELF.

Additionally, build the cosimulation executable with -O2 optimization to
prevent simulation execution timeouts during long-running integration
and RTOS tests.

* Align simulation test timeout overrides with cosim

Apply wildcard timeout overrides across all integration, RTOS,
and synchronization tests in run-all-tests so that multi-tasking
and interrupt-driven workloads do not hit default cycle limits
on parallel CI runner execution.

* Fix FreeRTOS and CSR trap decode on RV64

When running FreeRTOS preemptive multitasking on RV64, the idle task
failed to receive timer interrupts due to mstatus.MIE remaining zero
after returning from context switches. The microcode sequencer forces
mstatus CSR address 0x300 during MSTATUS_MRET, but the CPU CSR read
multiplexer decoded the 1-cycle registered csr_addr_reg (holding mepc
from the prior step) rather than the active combinational sequencer
address, overriding mstatus with mepc.

Decode the effective CSR address directly from useq_addr_out when the
microcode sequencer is active to eliminate the address registration
latency. Also adjust RV64 trap CSR WARL expectations in unit tests
and align CLINT store snoop bus widths with XLEN.

+++ testbench/lib/elf_loader.h
@@ -4,9 +4,9 @@
 #include <cstdint>
 #include <functional>
 
-// Load an ELF32 RISC-V binary into memory via a callback.
+// Load an ELF32 or ELF64 RISC-V binary into memory via a callback.
 // The callback is called for each word: callback(byte_addr, data_word).
 // Returns total bytes loaded, or -1 on error.
-int load_elf(const char* path, std::function<void(uint32_t, uint32_t)> write_word);
+int load_elf(const char* path, std::function<void(uint64_t, uint32_t)> write_word);
 
 #endif // ELF_LOADER_H
